+++ Specs/MockProperty.js
@@ -18,5 +18,12 @@ define(function() {
         return this.value;
     };
 
+    MockProperty.prototype.getValueRangeCartesian = function(start, stop, currentTime, result) {
+        this.lastStart = start;
+        this.lastStop = stop;
+        this.lastCurrentTime = currentTime;
+        return this.value;
+    };
+
     return MockProperty;
 });